requirements.txt
flask
flask_cors
flask_pymongo
asyncio
motor
requests
pymongo
langchain-community
langchain_core
langchain_google_genai
langgraph
argparse
flask-socketio
reportlab
langchain-mongodb
fpdf
bs4
pandas
sentence_transformers
